### 2018-06-28 #796 @whedon
[PRE REVIEW]: The hhi Package: Streamlined Calculation and Visualization of Herfindahl-Hirschman Index Scores
**Submitting author:** @pdwaggoner (<a href="http://orcid.org/0000-0002-7825-7573">Philip D. Waggoner</a>)
**Repository:** <a href="https://github.com/pdwaggoner/hhi" target ="_blank">https://github.com/pdwaggoner/hhi</a>
**Version:** v1.1.0
