### 2018-07-01 #796 @arfon
👋 @pdwaggoner. We'll need an editor to volunteer to handle this submission before we can get going. In the meantime could you take a look at addressing the compilation error? It looks like the YAML header for your paper is malformed.
