### 2018-07-02 #796 @leeper
@pdwaggoner Thanks for your submission. In order to review it, we'll need the repository to be in standard R package form (as in the form you build using `R CMD build` containing `/R`, `/man`, `/tests` directories) not isolated R code files.
