### 2018-07-04 46eca6f
Update paper.md
