### 2018-07-04 #796 @pdwaggoner
@arfon Thanks for this. I have updated the YAML header. Should be good to go now. Thanks for catching this. Also, @leeper thanks so much for serving as editor and for the guidance. I have included the standard R package form in my hhi repository. I think everything should be in there and good to go. Let me know if you need anything or else, or if you find another oversight. Many thanks.
