### 2018-07-04 #796 @leeper
@pdwaggoner To clarify, the repo should contain all the code such that it is visible and navigable in the GitHub interface and can be installed and built by the reviewer, in standard package format. We don't need the .tar.gz file.

As an example, see the structure of the repository for this recent JOSS publication: https://github.com/strengejacke/ggeffects
