### 2018-07-11 #796 @leeper
@strengejacke Would you be willing to review this for JOSS?
