### 2018-07-20 #796 @pdwaggoner
@leeper I merged pull request. All should be good. Please excuse the .bib oversight
