### 2018-07-20 fa33a42
Update paper.md
