### 2018-07-20 #796 @pdwaggoner
@leeper Not sure if you are looking for a response from me re: this proof, but this looks good to me. Let me know whatever else you need, if anything. Thanks!
