### 2018-07-20 #828 @leeper
@cskovron The review issue is here. Just work through the checklist (above), [open an issue here](https://github.com/pdwaggoner/hhi/issues) for anything missing or any other changes you'd like to see, and then ping me here when you've completed your review. Let me know if you have any questions.
