### 2018-07-28 85f8a5a
Update footwear example.R
