### 2018-07-28 0fbb52e
Update footwear example.R
