### 2018-08-02 #828 @leeper
@cskovron Just a nudge on this review.
