### 2018-08-03 #828 @leeper
Thanks, @cskovron! Appreciate the quick review!

@pdwaggoner I have a few additional points I'd like to see addressed:

- [ ] You'll need to change the LICENSE file to be LICENSE.md or LICENSE.txt and add it to an .Rbuildignore file.
- [ ] For LICENSE, just have your name and copyright dates (like this: https://github.com/tidyverse/dplyr/blob/master/LICENSE), as this is what CRAN expects for an MIT-licensed package
- [ ] Add the following files to .Rbuildignore:
  - LICENSE.md/LICENSE.txt (as above)
  - hhi_manual.pdf
  - microwaves.csv
  - paper.bib
  - paper.md
  - plot1.png
  - plot2.png
- [ ] Please add a `/tests` directory that contains an automated test (e.g., using testthat) that verifies the calculated index is correct for either a simulated or example dataset. More information on testing can be found in: http://r-pkgs.had.co.nz/tests.html. An example of what I'm looking for can be found here: https://github.com/leeper/GK2011/blob/master/tests/testthat/tests.R.
- [ ] I believe you can simplify these lines: https://github.com/pdwaggoner/hhi/blob/master/R/hhi.R#L31-L37 to just be the following vectorized alternative, which will be faster:

```R
if (any(d < 0)) {
    stop('vector "s" must contain only positive values')
}
hhi <- sum(d^2)
```

- [ ] Optional: consider setting up a continuous integration service. For details: http://r-pkgs.had.co.nz/check.html

Once that's done, please deposit a copy of your software repository (including any revisions made during the JOSS review process) with a data-archiving service. To do so:

 3. Post a comment here to @leeper with the DOI for the release.

Let me know if you have any questions about the process.
