### 2018-08-03 de85e05
Create LICENSE.txt
