### 2018-08-03 #828 @pdwaggoner
@leeper This sounds great. I will get on these things in a day or so. Thanks. To be extra clear, did you want me to resubmit my package to CRAN after adding these things, or did you want me to make the changes only in GitHub (e.g., in the current release on CRAN, the LICENSE file is currently a .txt in the exact form you suggested; its only more verbose on GitHub)? Just want to get it right the first time. Thanks!
