### 2018-08-04 #828 @leeper
For JOSS, I'm only concerned about GitHub. CRAN release is up to you.
