### 2018-08-05 b75442f
Update hhi.R
