+++ R/hhi.R
@@ -28,12 +28,10 @@ hhi <- function(x, s){
     stop('"s" must be numeric vector\n',
          'You have provided an object of class: ', class(d)[1])
   }
-  for(i in 1:length(d)) {
-    if(d[i] < 0){
-      stop('vector "s" must contain only positive values')
-    }
-    d[i] <- d[i]^2
-    hhi <- sum(d)
-  }
+  if (any(d < 0)) {
+    stop('vector "s" must contain only positive values')
+}
+hhi <- sum(d^2)
+  
   return(hhi)
 }
