### 2018-08-05 b4e8589
Update hhi_test.R
