+++ README.md
@@ -1,6 +1,8 @@
 # `hhi` An R Package to Calculate and Visualize the Herfindahl-Hirschman Index
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/hhi)](http://cran.r-project.org/package=hhi)
 [![total](http://cranlogs.r-pkg.org/badges/grand-total/hhi)](http://cranlogs.r-pkg.org/)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1336788.svg)](https://doi.org/10.5281/zenodo.1336788)
+
 
 ## Why should I use `hhi`?
 
