### 2018-08-05 #828 @pdwaggoner
@leeper Alright - I think everything is good to go. I followed all steps, and think the tests should be structure properly as well. I created the release and also generated the DOI with Zenodo. Here is the target URL they provided: https://doi.org/10.5281/zenodo.1336788 . Let me know what next steps are. I would like to quickly update the sample code structure in my paper prior to finalizing it with JOSS and also update my thanks section, if at all possible. Thanks again!
