### 2018-08-05 #828 @leeper
@pdwaggoner Great. Please update your paper accordingly and we'll regenerate the PDF.
