### 2018-08-05 bdff883
Update paper.md
