### 2018-08-05 c008bdc
Update paper.md
