+++ paper.md
@@ -35,12 +35,12 @@ microwaves <- read.csv(".../microwaves.csv") # Specify file path where .csv file
 
 # Next, store all HHI scores for each year in the data
 # Note, for several years a warning message notes shares don't sum to 100 due to rounding
-usa.12 <- hhi(usa, "ms.2012")<br/>
-usa.13 <- hhi(usa, "ms.2013")<br/>
-usa.14 <- hhi(usa, "ms.2014")<br/>
-usa.15 <- hhi(usa, "ms.2015")<br/>
-usa.16 <- hhi(usa, "ms.2016")<br/>
-usa.17 <- hhi(usa, "ms.2017")<br/>
+usa.12 <- hhi(usa, "ms.2012")
+usa.13 <- hhi(usa, "ms.2013")
+usa.14 <- hhi(usa, "ms.2014")
+usa.15 <- hhi(usa, "ms.2015")
+usa.16 <- hhi(usa, "ms.2016")
+usa.17 <- hhi(usa, "ms.2017")
 
 # Create a few objects to see fluctuation in the U.S. HHI from 2012-2017
 usa.hhi <- rbind(usa.12, usa.13, usa.14, usa.15, usa.16, usa.17)
