### 2018-08-05 899a942
Update paper.md
