### 2018-08-06 e7ddd86
tweak code chunks in paper.md for JOSS
