### 2018-08-06 #828 @leeper
@pdwaggoner Great. I also just sent you a PR: https://github.com/pdwaggoner/hhi/pull/3, which will hopefully get the code chunks rendering correctly in the PDF.
