### 2018-08-06 #828 @pdwaggoner
@leeper Perfect; its merged. I was wondering why it did that, but I suppose it was just an old format for inserting code into md. Many thanks for the find and fix!
