### 2018-08-06 #828 @leeper
@pdwaggoner Thanks. Take a look at the PDF and make sure everything looks ok. Seems like you might want to tweak the paper.md to wrap some of the longer code and comment lines. That's a purely aesthetic change, though.
