### 2018-08-06 #828 @pdwaggoner
Thanks @leeper . I updated the code a bit and it looks fine in my repo. Would you mind generating one more time to make sure it looks alright when compiling the PDF?
